% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melviewr.R
\name{melviewr}
\alias{melviewr}
\title{melviewr}
\usage{
melviewr(melodic_dir, standard_file = NULL, motion_file = NULL)
}
\arguments{
\item{melodic_dir}{string Path to MELODIC output directory}

\item{standard_file}{string Optional path to a 3-dimensional Nifti standard file
of the same voxel dimensions as the melodic output}

\item{motion_file}{string Optional path to a summary motion text file. This file
should have one column and as many rows as there are volumes in the functional
data}
}
\value{
NULL The function opens a GUI.
}
\description{
View and Classify Components from a Melodic Analysis
}

