% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melviewr.R
\docType{class}
\name{Viewr-class}
\alias{Viewr}
\alias{Viewr-class}
\title{A Reference Class that represents the melviewr GUI and all its data}
\description{
A Reference Class that represents the melviewr GUI and all its data
}
\section{Fields}{

\describe{
\item{\code{win}}{An instance of a gwidgets gwindow.}

\item{\code{widgets}}{A list to hold gwidgets widgets.}

\item{\code{settings}}{A list to hold graphics settings and defaults.}

\item{\code{status}}{A list to hold state-related status information.}

\item{\code{data}}{A list to hold all necessary data the GUI uses.}
}}

